Added docs, moved some stuff around, created an interceptor to use instead of grabbing the token before each request
